+++ client/CHANGELOG.md
@@ -8,3 +8,4 @@
 - Created PrivateRoute component to protect routes that require authentication. Redirects to login page if user is not authenticated
 - Session token is stored in local storage. auth.ts contains function isAuth to check if user is authenticated. Function currently only checks if storage contains token. In the future it will need to check if token is valid with the server.
 - Majorly overrode the login page and installed react-bootstrap for styling components
+- Replaced the old header component with a new one using the react-bootstrap navbar component